refactor(lib): split up types.rs into multiple files
