+++ test-suite/src/formatting.rs
@@ -5,7 +5,7 @@ mod test {
     use crate::test_helpers::{cargo_dot_toml, NON_RESPONSE_NUM};
     use lspresso_shot::{
         lspresso_shot, test_formatting,
-        types::{FormattingResult, ServerStartType, TestCase, TestError, TestFile},
+        types::{formatting::FormattingResult, ServerStartType, TestCase, TestError, TestFile},
     };
     use test_server::{get_dummy_server_path, send_capabiltiies, send_response_num};
